Features added (Tests): Fleshing out FSM
* SM_PERSISTABLE: Removed
* MOCK_DATA_ONE: Renamed to MOCK_DOOR
* MOCK_DOOR: Added features in an attempt to create a managed object
* MOCK_MACHINE: Added features in an attempt to manage MOCK_DOOR
* Tests: Built tests to help expose bugs in design -- YAY! We have bugs!

NOTE-1: Commented out code that is buggy, but let some in. The machine
is working as a mere pub-sub, but it is not quite there in terms of
being an implemented FSM.

NOTE-2: It is clear that from the view of the Client, attempting to set
up an FSM with a managed object (like MOCK_DOOR), that there are design
issues and API issues.